requirements.txt
alembic==1.8.0
amqp==5.1.1
aniso8601==9.0.1
bcrypt==3.2.2
billiard==3.6.4.0
blinker==1.4
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.0
Flask-Mail==0.9.1
Flask-Mail-SendGrid==0.3
Flask-Migrate==3.1.0
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.0
MarkupSafe==2.1.1
mongoengine==0.24.1
Pillow==9.1.1
prompt-toolkit==3.0.30
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pymongo==4.1.1
python-dotenv==0.20.0
python-http-client==3.3.7
pytz==2022.1
sendgrid==6.9.7
six==1.16.0
SQLAlchemy==1.4.36
starkbank-ecdsa==2.0.3
urllib3==1.26.9
validators==0.19.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
